requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pillow==10.2.0
torch==2.3.0
torchvision==0.18.0
numpy==1.26.4